backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
numpy
groq
python-dotenv